+++ opentelemetry-sdk/src/opentelemetry/sdk/util/__init__.py
@@ -12,6 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import copy
 import datetime
 import threading
 from collections import deque
@@ -54,6 +55,14 @@ class BoundedList(Sequence):
         self.dropped = 0
         self._dq = deque(maxlen=maxlen)  # type: deque
         self._lock = threading.Lock()
+
+    def __deepcopy__(self, memo):
+        copy_ = BoundedList(0)
+        memo[id(self)] = copy_
+        with self._lock:
+            copy_.dropped = self.dropped
+            copy_._dq = copy.deepcopy(self._dq, memo)
+        return copy_
 
     def __repr__(self):
         return f"{type(self).__name__}({list(self._dq)}, maxlen={self._dq.maxlen})"